Remove cursor() and controller() from MathQuill namespace in mathquill.d.ts (#24)
## Summary:
There was a lot of weirdness with trying to add the
types for cursor() and controller() here. It either
breaks all the types in MathQuill, or it breaks all
the types in Perseus. Alternatively, we were going
to [add the types in the build/ folder but not in src/](https://github.com/Khan/mathquill/pull/19),
but then we wouldn't be able to keep those two
mathquill.d.ts files in sync.

The new approach we're trying is to go back to what
this was before in mathquill - not having these types
defined here - and instead, define these types on top
in Perseus. This should remove all the type errors
and still allow us to use the types the way we want.

Issue: https://khanacademy.atlassian.net/browse/LC-1659

## Test plan:
`make lint`

Author: nishasy

Reviewers: nedredmond

Required Reviewers:

Approved By: nedredmond

Checks: ✅ build (20.x), ✅ build (18.x), ✅ build (16.x)

Pull Request URL: https://github.com/Khan/mathquill/pull/24