Change font references to relative paths (#30)
## Summary:

This PR adjusts the Symbola font references in Mathquill Less files to be relative paths. It also updates the build pipeline so that the generated CSS contains correct relative paths to these font files relative to the build output folder. 

I also noticed that there was a reference to an `.otf` font file that no longer exists in the repo. It seems like this is just "code rot" as the code changed.

Issue: LEMS-3019

## Test plan:

I ran the build and inspected the CSS and made sure that the relative font paths for Symbola resolved to the actual files in the build folder.

Author: jeremywiebe

Reviewers: jeremywiebe, jeresig, nedredmond, benchristel

Required Reviewers:

Approved By: jeresig, nedredmond

Checks: ✅ 3 checks were successful

Pull Request URL: https://github.com/Khan/mathquill/pull/30